% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_distribution.R
\name{dA}
\alias{dA}
\title{Density function of the Conjugate DP distribution}
\usage{
dA(x, a, b, size, log = FALSE)
}
\arguments{
\item{x}{Vector of values to evaluate}

\item{a}{Parameter exponentiating the numerator. It must hold that \code{a > 0} and \code{1 < a/b < size}}

\item{b}{Parameter exponentiating the denominator. Default is \code{b = 1}. It must hold that \code{b > 0} and  \code{1 < a/b < size}}

\item{size}{Size of the sample form a Dirichlet process. Must be an integer greater or equal than 2}

\item{log}{return the density in log scale}
}
\value{
A vector of size \code{length(x)}
}
\description{
Density function of the Conjugate DP distribution
}
