% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stirling_gamma.R
\name{dStirgamma}
\alias{dStirgamma}
\title{Density function of the Conjugate DP distribution}
\usage{
dStirgamma(x, a, b, m, log = FALSE)
}
\arguments{
\item{x}{Vector of values to evaluate}

\item{a}{Location parameter of the Stirling-gamma. It must hold that \code{a > 0} and \code{1 < a/b < m}.}

\item{b}{Precision parameter of the Stirling-gamma. Default is \code{b = 1}. It must hold that \code{b > 0} and \code{1 < a/b < m}.}

\item{m}{Pseudo sample size parameter. Must be an integer greater or equal than 2}

\item{log}{return the density in log scale}
}
\value{
A vector of size \code{length(x)}
}
\description{
Density function of the Conjugate DP distribution
}
