% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_distribution.R
\name{rA}
\alias{rA}
\title{Random generator for the A}
\usage{
rA(n = 1, a, b = 1, size)
}
\arguments{
\item{n}{Number of samples to take. Default is \code{n = 1}}

\item{a}{Parameter exponentiating the numerator. It must hold that \code{a > 0} and \code{1 < a/b < size}}

\item{b}{Parameter exponentiating the denominator. Default is \code{b = 1}. It must hold that \code{b > 0} and  \code{1 < a/b < size}}

\item{size}{Size of the sample form a Dirichlet process. Must be an integer greater or equal than 2}
}
\value{
A vector of size \code{n}
}
\description{
Random generator for the A
}
\examples{

n <- 1000
size <- 100
a <- 30
b <- 1
samples <- rA(n, a, b, size)
}
