% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonConjugate_Stirling_gamma.R
\name{rNCStirgamma}
\alias{rNCStirgamma}
\title{Sampler for the posterior distribution for the precision parameter of the Dirichlet
process under the NonConjugate Stirling-gamma prior.}
\usage{
rNCStirgamma(nsamples = 1, a, b = 1, m, n, k)
}
\arguments{
\item{nsamples}{Number of samples to take. Default is \code{nsamples = 1}}

\item{a}{Location parameter of the Stirling-gamma. It must hold that \code{a > 0} and \code{1 < a/b < m}.}

\item{b}{Precision parameter of the Stirling-gamma. Default is \code{b = 1}. It must hold that \code{b > 0} and \code{1 < a/b < m}.}

\item{m}{Pseudo sample size parameter. Must be an integer greater or equal than 2}

\item{n}{Observed sample size, Must be an integer greater or equal than 2}

\item{k}{Number of distinct clusters observed in \code{n} observations. Must be an integer greater than 2.}
}
\value{
A vector of size \code{nsamples}
}
\description{
Sampler for the posterior distribution for the precision parameter of the Dirichlet
process under the NonConjugate Stirling-gamma prior.
}
\examples{
nsamples <- 1000
m <- 100
a <- 30
b <- 1
k <- 10
n <- 500
samples <- rNCStirgamma(nsamples, a, b, m, n, k)
}
