% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSg.R
\name{rSg}
\alias{rSg}
\title{Random number generator for the Stirling-gamma distribution}
\usage{
rSg(nsamples = 1, a, b, m, method = NULL)
}
\arguments{
\item{nsamples}{Number of samples to take. Default is \code{nsamples = 1}}

\item{a}{Location parameter of the Stirling-gamma. It must hold that \code{a > 0} and \code{1 < a/b < m}.}

\item{b}{Precision parameter of the Stirling-gamma. It must hold that \code{b > 0} and \code{1 < a/b < m}.}

\item{m}{Reference sample size parameter. Must be an integer, so that \code{1 < a/b < m}.}

\item{method}{Sampling algorithm method for the Stirling-gamma. Available options are
\code{'beta_prime'} and \code{'ratio_uniform'}. Default value is \code{NULL}, which sets
\code{method = 'beta_prime'} if \code{a - b <= 1}, and \code{method = 'ratio_uniform'} when \code{a - b > 1}.}
}
\value{
A vector of size \code{nsamples}
}
\description{
Random number generator for the Stirling-gamma distribution
}
\examples{

nsamples <- 1000
m <- 100
a <- 5
b <- 1
samples <- rSg(nsamples, a, b, m)
}
