% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSg_posterior.R
\name{rSg_posterior}
\alias{rSg_posterior}
\title{Random number generator for the updated Stirling-gamma distribution}
\usage{
rSg_posterior(nsamples = 1, a, b, m, k, n)
}
\arguments{
\item{nsamples}{Number of samples to take. Default is \code{nsamples = 1}}

\item{a}{Location parameter of the Stirling-gamma. It must hold that \code{a > 0} and \code{1 < a/b < m}.}

\item{b}{Precision parameter of the Stirling-gamma. It must hold that \code{b > 0} and \code{1 < a/b < m}.}

\item{m}{Reference sample size parameter. Must be an integer, so that \code{1 < a/b < m}.}

\item{k}{Number of clusters detected in \code{n} observations. Must be \code{1 <= k <= n}.}

\item{n}{Number of observations. Must be an integer \code{n >= 1}.}
}
\value{
A vector of size \code{nsamples}
}
\description{
Random number generator for the updated Stirling-gamma distribution
}
\examples{

nsamples <- 1000
m <- 10
a <- 3
b <- 1
n <- 100
k <- 5
samples <- rSg_posterior(nsamples, a, b, m, k, n)
}
