% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stirling_gamma.R
\name{rStirgamma}
\alias{rStirgamma}
\title{Random number generator for the Stirling-gamma distribution}
\usage{
rStirgamma(nsamples = 1, a, b = 1, m)
}
\arguments{
\item{nsamples}{Number of samples to take. Default is \code{nsamples = 1}}

\item{a}{Location parameter of the Stirling-gamma. It must hold that \code{a > 0} and \code{1 < a/b < m}.}

\item{b}{Precision parameter of the Stirling-gamma. Default is \code{b = 1}. It must hold that \code{b > 0} and \code{1 < a/b < m}.}

\item{m}{Pseudo sample size parameter. Must be an integer greater or equal than 2}
}
\value{
A vector of size \code{nsamples}
}
\description{
Random number generator for the Stirling-gamma distribution
}
\examples{

nsamples <- 1000
m <- 100
a <- 30
b <- 1
samples <- rStirgamma(nsamples, a, b, m)
}
